requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.1
attrs==25.3.0
autobahn==25.10.2
Automat==25.4.16
beautifulsoup4==4.13.5
billiard==4.2.2
bleach==6.2.0
boto3==1.40.45
botocore==1.40.45
cachetools==6.2.2
celery==5.3.1
certifi==2025.11.12
cffi==2.0.0
channels==4.3.1
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
config==0.5.1
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
dataclasses-json==0.6.7
decorator==4.4.2
defusedxml==0.7.1
distro==1.9.0
Django==5.2.5
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docx2txt==0.9
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.9.1
elevenlabs==2.23.0
fastjsonschema==2.21.2
frozenlist==1.7.0
git-filter-repo==2.47.0
google-auth==2.43.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
hyperlink==21.0.0
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
incremental==24.7.2
inflection==0.5.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kombu==5.5.4
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-pinecone==0.2.12
langchain-text-splitters==0.3.11
langsmith==0.4.27
lxml==6.0.2
MarkupSafe==3.0.2
marshmallow==3.26.1
mistune==3.1.4
moviepy==1.0.3
multidict==6.6.4
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.3.3
openai==1.107.1
orjson==3.11.3
packaging==24.2
pandocfilters==1.5.1
pdf2image==1.17.0
pillow==11.3.0
pinecone==7.3.0
pinecone-client==6.0.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
platformdirs==4.4.0
proglog==0.1.12
prompt_toolkit==3.0.52
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.10.1
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pypdf==6.0.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.1.1
PyYAML==6.0.2
pyzmq==27.1.0
redis==4.5.5
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
service-identity==24.2.0
setuptools==80.9.0
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sqlparse==0.5.3
stripe==13.2.0
tenacity==9.1.2
tiktoken==0.11.0
tinycss2==1.4.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
Twisted==25.5.0
txaio==25.9.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
webencodings==0.5.1
websockets==15.0.1
wheel==0.45.1
yarl==1.20.1
zope.interface==8.1.1
zstandard==0.24.0
